test(KEN-801): the refresh control gets its own file [no-changelog]
The baseline row is gone. It pinned package.test.tsx at 874 against an
800-line class permanently, for every future edit to that file, which is not
an accommodation the cut asked for.

package.test.tsx is main's content again, at 712 lines. The control lives in
package-refresh.test.tsx with the fixture it needs and nothing else: the
scan and nav state for one package in one place, the stubs the page calls on
mount, and the three commands the card's Update runs behind it. No shared
harness comes back.

The ratchet passes with no baseline row and no declaration.
