+++ ui/src/pages/package-refresh.test.tsx
@@ -1,0 +1,276 @@
+// @vitest-environment jsdom
+// What an update started from the package page leaves behind. The Projects
+// tab's Update commits through the updates store, which refreshes the scan
+// and the audit and knows nothing about this page's own three reads: its
+// card was made to follow the landed commit in #1799, while the Overview and
+// the header went on describing the copy the update replaced.
+//
+// One control, and the fixture it needs. The page's other tests are
+// `package.test.tsx`'s.
+import { act } from "react";
+import { beforeEach, describe, expect, it, vi } from "vitest";
+import type {
+  AuditView_Serialize,
+  ObservedItem,
+  ScanResult,
+  Scope,
+  UpdateRow,
+  VersionRow,
+} from "@/bindings";
+import { commands } from "@/bindings";
+import { ADOPTABLE } from "@/lib/adoptable";
+import { UPDATE_LABEL } from "@/lib/copy";
+import { OVERVIEW_TAB } from "@/lib/copy-customize";
+import { PROJECTS_TAB, updateInLabel } from "@/lib/copy-projects";
+import { READ_LANDED } from "@/lib/read-state";
+import { useAuditStore } from "@/stores/audit";
+import { useNavStore } from "@/stores/nav";
+import { useScanStore } from "@/stores/scan";
+import { useUpdatesStore } from "@/stores/updates";
+import { mount, settle } from "@/test/dom";
+import { PackagePage } from "./package";
+
+// The page is mounted against the real stores; only the backend is stubbed.
+vi.mock("@/bindings", async (importOriginal) => ({
+  ...(await importOriginal<typeof import("@/bindings")>()),
+  commands: {
+    packageMeta: vi.fn(),
+    packageFiles: vi.fn(),
+    packageVersions: vi.fn(),
+    packageReadme: vi.fn(),
+    getManifest: vi.fn(),
+    editorInventory: vi.fn(),
+    getScopeSettings: vi.fn(),
+    revealPath: vi.fn(),
+    openInEditor: vi.fn(),
+    libraryProvenance: vi.fn(),
+    packageDiff: vi.fn(),
+    auditAll: vi.fn(),
+    // What the card's Update runs, and the two reads the store lands behind
+    // it.
+    packageUpdate: vi.fn(),
+    updatesOverview: vi.fn(),
+    scanMachine: vi.fn(),
+  },
+}));
+
+const VG: Scope = { scope: "project", root: "/work/vg" };
+const OLD = "a".repeat(40);
+const NEW = "b".repeat(40);
+
+const nothing = { status: "error" as const, error: "not in this test" };
+
+/** gh as the scan found it, in the one place this page is about. */
+const INSTALLED: ObservedItem = {
+  kind: "skill",
+  name: "gh",
+  scope: VG,
+  harness: "claude",
+  path: "/work/vg/.claude/skills/gh",
+  fileState: { state: "file" },
+  enabled: true,
+  origin: null,
+  description: "about gh",
+  tags: [],
+  modifiedAt: null,
+  vendor: null,
+};
+
+/** This place's row: the commit installed there, and whether the check found
+ *  something newer waiting for it. */
+const rowAt = (commit: string, waiting: boolean): UpdateRow => ({
+  scope: VG,
+  kind: "skill",
+  name: "gh",
+  source: "cat",
+  repo: "o/r",
+  repoIdentity: "o/r",
+  current: { commit, label: null, date: null },
+  latest: waiting ? { commit: NEW, label: null, date: null } : null,
+  updateAvailable: waiting,
+  pinned: false,
+  holdOwner: null,
+  ignored: false,
+  blockedByLocalEdit: false,
+  editedHarnesses: [],
+  forkableHarness: null,
+  canDiscard: false,
+  canTakeLatest: false,
+  derived: false,
+  requiredBy: [],
+  forked: false,
+  mixed: false,
+  removedUpstream: false,
+  noPerPackageUpdate: null,
+});
+
+const version = (
+  id: string,
+  label: string,
+  installed: boolean,
+): VersionRow => ({
+  id,
+  label,
+  date: "2026-08-28T12:00:00Z",
+  summary: "release notes",
+  installed,
+  newerThanInstalled: !installed,
+});
+
+/** The scope view an apply answers with: it wrote, and there is nothing else
+ *  to say about the place. */
+const APPLIED: AuditView_Serialize = {
+  scope: VG,
+  drift: [],
+  plan: [],
+  notes: [],
+  warnings: [],
+  safety: [],
+  adoptable: ADOPTABLE,
+  exits: [],
+};
+
+beforeEach(() => {
+  vi.clearAllMocks();
+  vi.mocked(commands.auditAll).mockResolvedValue({ status: "ok", data: [] });
+  vi.mocked(commands.packageReadme).mockResolvedValue(nothing);
+  vi.mocked(commands.editorInventory).mockResolvedValue(nothing);
+  vi.mocked(commands.libraryProvenance).mockResolvedValue(nothing);
+  vi.mocked(commands.packageDiff).mockResolvedValue(nothing);
+  vi.mocked(commands.getManifest).mockResolvedValue({
+    status: "ok",
+    data: { manifest: { schema: 1, install: {} }, base: null },
+  });
+  vi.mocked(commands.getScopeSettings).mockResolvedValue({
+    status: "ok",
+    data: { applies: true, skills: [], base: null },
+  });
+  useAuditStore.setState({ views: [], auditedAt: null, read: READ_LANDED });
+});
+
+/** The page open on gh in vg, with the scan holding the one installation. */
+const openPage = async () => {
+  useScanStore.setState({
+    result: {
+      harnesses: [],
+      items: [INSTALLED],
+      missingProjects: [],
+      warnings: [],
+    },
+  });
+  useNavStore.setState({
+    page: "package",
+    packageRef: { kind: "skill", name: "gh", scope: VG },
+    packageView: null,
+  });
+  const host = mount(<PackagePage />);
+  await settle();
+  return host;
+};
+
+const header = (host: HTMLElement) => host.querySelector("header")?.textContent;
+
+const openTab = async (host: HTMLElement, name: string) => {
+  const found = [...host.querySelectorAll('[data-slot="tabs-trigger"]')].find(
+    (trigger) => trigger.textContent === name,
+  );
+  if (!found) throw new Error(`no ${name} tab`);
+  await act(async () => {
+    (found as HTMLElement).click();
+  });
+  await settle();
+};
+
+describe("the package page after an update started from its Projects tab", () => {
+  it("re-reads its files, its version and its update offer", async () => {
+    // What the engine answers before and after the apply lands: core stamps
+    // the whole installation when the source hash moves, so the record, the
+    // timeline and the files move together.
+    let landed = false;
+    vi.mocked(commands.packageMeta).mockResolvedValue({
+      status: "ok",
+      data: {
+        source: "cat",
+        repo: "o/r",
+        repoUrl: null,
+        rev: null,
+        current: null,
+        installedAt: null,
+        harnesses: ["claude"],
+        enabled: true,
+        fork: null,
+        catalog: null,
+      },
+    });
+    vi.mocked(commands.packageVersions).mockImplementation(() =>
+      Promise.resolve({
+        status: "ok",
+        data: landed
+          ? [version(NEW, "v2", true)]
+          : [version(NEW, "v2", false), version(OLD, "v1", true)],
+      }),
+    );
+    vi.mocked(commands.packageFiles).mockImplementation(() =>
+      Promise.resolve({
+        status: "ok",
+        data: [
+          {
+            path: landed ? "AFTER.md" : "BEFORE.md",
+            size: 10,
+            isReadme: false,
+          },
+        ],
+      }),
+    );
+    vi.mocked(commands.packageUpdate).mockImplementation(() => {
+      landed = true;
+      return Promise.resolve({
+        status: "ok",
+        data: { view: APPLIED, heldBack: [], removed: [], moved: [] },
+      });
+    });
+    // The standing read the store lands behind its own apply.
+    vi.mocked(commands.updatesOverview).mockImplementation(() =>
+      Promise.resolve({
+        status: "ok",
+        data: {
+          rows: [rowAt(landed ? NEW : OLD, !landed)],
+          warnings: [],
+          unreadable: [],
+          lastFetched: null,
+        },
+      }),
+    );
+    // The rescan behind the apply finds the same machine: the package is
+    // installed where it was, so the page stays on screen.
+    vi.mocked(commands.scanMachine).mockImplementation(() =>
+      Promise.resolve({
+        status: "ok",
+        data: useScanStore.getState().result as ScanResult,
+      }),
+    );
+    useUpdatesStore.setState({ rows: [rowAt(OLD, true)], read: READ_LANDED });
+
+    const host = await openPage();
+    expect(host.textContent).toContain("BEFORE.md");
+    expect(host.textContent).toContain("v1");
+    expect(header(host)).toContain(UPDATE_LABEL);
+
+    await openTab(host, PROJECTS_TAB);
+    const update = [...host.querySelectorAll("button")].find(
+      (one) => one.getAttribute("aria-label") === updateInLabel("vg"),
+    );
+    if (!update) throw new Error("no Update on the vg card");
+    await act(async () => {
+      update.click();
+    });
+    await settle();
+    await openTab(host, OVERVIEW_TAB);
+
+    expect(host.textContent).toContain("AFTER.md");
+    expect(host.textContent).not.toContain("BEFORE.md");
+    expect(host.textContent).toContain("v2");
+    expect(host.textContent).not.toContain("v1");
+    expect(header(host)).not.toContain(UPDATE_LABEL);
+  });
+});
